PSMDB-135 adapt parser code to rocksdb 5.7.3
and add support for multiple rocksdb versions (based on preprocessor conditions)

(cherry picked from commit 5aad3f1c81fb6cb27fe28a525b40b1db50ee8739)
